Tidy from vincular code
I've created two new functions on tiling, one creates a vincular pattern or bivincular pattern from a cperm (or a string of a cperm) and input of adjacent indices and values using Reed's existing code, but removes the empty rows/columns rather than making them empty so that they don't have to be tidied in the first step for mapplings. I've also added code that creates empty cells or regions on a tiling for making mesh patterns. They are both intended to make running the mapplings code easier for users (and us!), let me know what you think and if it's useful. Reed's old code is still there too.